:recycle: Refactor: 폴더, 파일, 변수명 통일(#142)
